# openapi.json
{"openapi":"3.0.1","info":{"title":"Non Obstacle Tourist Swagger","description":"Non Obstacle Tourist Swagger","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"회원관리 API","description":"회원가입, 로그아웃, Access Token 재발급 관련 API입니다."},{"name":"회원관리 API","description":"회원가입, 로그아웃, Access Token 재발급 관련 API입니다. "},{"name":"관광지 DB 초기화 API","description":"공공 API를 통한 로컬 DB 초기화에 대한 설명입니다."},{"name":"관광지 API","description":"관광지 API에 대한 설명입니다."},{"name":"검색 API","description":"각종 검색 API입니다."},{"name":"공지사항 API","description":"공지사항 API입니다."},{"name":"마이페이지 API","description":"마이페이지 관련 API입니다."}],"paths":{"/api/v1/likes/{touristAttractionId}":{"put":{"tags":["회원관리 API"],"summary":"관광지 좋아요 업데이트 API","description":"관광지 좋아요 업데이트 API 입니다.","operationId":"updateTouristAttractionLikes","parameters":[{"name":"touristAttractionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TouristAttractionLikeResponseDto"}}}}}},"delete":{"tags":["회원관리 API"],"summary":"관광지 좋아요 삭제 API","description":"관광지 좋아요 삭제 API 입니다.","operationId":"deleteTouristAttractionLikes","parameters":[{"name":"touristAttractionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/join":{"post":{"tags":["회원관리 API"],"summary":"회원가입 API","description":"회원가입 API입니다.","operationId":"join","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SignUpResponseDto"}}}}}}},"/api/v1/notices":{"get":{"tags":["공지사항 API"],"summary":"공지사항 리스트 API","description":"공지사항 리스트 API 입니다.","operationId":"getNotices","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NoticeDetailResponseDto"}}}}}}},"post":{"tags":["공지사항 API"],"summary":"공지사항 추가 API ","description":"공지사항 추가 API 입니다.","operationId":"addNotice","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoticeRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NoticeResponseDto"}}}}}}},"/api/v1/mypage/inquiries":{"get":{"tags":["마이페이지 API"],"summary":"1대1 문의 리스트 API","description":"1대1 문의 리스트 API 입니다.","operationId":"getInquiries","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":5}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InquiryDetailResponseDto"}}}}}}},"post":{"tags":["마이페이지 API"],"summary":"1대1 문의 추가 API","description":"1대1 문의 추가 API 입니다.","operationId":"addInquiry","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InquiryRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InquiryResponseDto"}}}}}}},"/api/v1/mypage/faqs":{"get":{"tags":["마이페이지 API"],"summary":"자주 묻는 질문 리스트 API","description":"자주 묻는 질문 리스트 API 입니다.","operationId":"getFaqs","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":5}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FaqDetailResponseDto"}}}}}}},"post":{"tags":["마이페이지 API"],"summary":"자주 묻는 질문 리스트 추가 API","description":"자주 묻는 질문 리스트 추가 API입니다.","operationId":"addFaq","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FaqRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FaqResponseDto"}}}}}}},"/api/v1/likes/touristAttraction":{"get":{"tags":["회원관리 API"],"summary":"관광지 좋아요 전체 조회 API","description":"관광지 좋아요 전체 조회 API 입니다.","operationId":"getTouristAttractionLikes_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TouristAttractionLikeResponseDto"}}}}}}},"post":{"tags":["회원관리 API"],"summary":"관광지 좋아요 추가 API","description":"관광지 좋아요 추가 API 입니다.","operationId":"addTouristAttractionLike","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TouristAttractionLikeRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TouristAttractionLikeResponseDto"}}}}}}},"/api/v1/likes/course":{"post":{"tags":["회원관리 API"],"summary":"코스 좋아요 추가 API","description":"코스 좋아요 추가 API 입니다.","operationId":"addCourseLike","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CourseLikeRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CourseLikeResponseDto"}}}}}}},"/refresh":{"get":{"tags":["회원관리 API"],"summary":"Access Token 재발급 API","description":"유효한 Refresh Token을 받고 Access Token을 재발급합니다.","operationId":"refresh","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/query":{"get":{"tags":["회원관리 API"],"summary":"로그인 테스트 API","description":"로그인 테스트 API입니다. 로그인 성공 시 query를 리턴합니다.","operationId":"query","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/logout":{"get":{"tags":["회원관리 API"],"summary":"관광지 에러 재시도 초기화","description":"하드코딩으로 저장 실패한 요소를 다시 시도합니다. 재빌드 하세요.","operationId":"logout","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/{areaCode}/{numOfRows}/{pageNo}":{"get":{"tags":["관광지 DB 초기화 API"],"summary":"단위 관광지 초기화 ","description":"Path Variable로 받은 관광지의 정보대로 공공API로부터 초기화시킵니다.","operationId":"initTouristAttraction","parameters":[{"name":"areaCode","in":"path","required":true,"schema":{"type":"string"}},{"name":"numOfRows","in":"path","required":true,"schema":{"type":"string"}},{"name":"pageNo","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/search/tourist-attraction":{"get":{"tags":["검색 API"],"summary":"관광지 검색 API","description":"지역,유형,검색어를 기반으로 검색합니다.","operationId":"searchTouristAttractions","parameters":[{"name":"areacode","in":"query","required":false,"schema":{"type":"string"}},{"name":"type","in":"query","required":false,"schema":{"type":"string"}},{"name":"searchParam","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":12}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseDtoListTouristSearchResponseDto"}}}}}}},"/api/v1/search/non-obstacle":{"get":{"tags":["검색 API"],"summary":"무장애 기반 관광지 검색 API","description":"무장애 기반 관광지 검색 API 입니다.","operationId":"searchTouristAttractionsByType","parameters":[{"name":"type","in":"query","required":true,"schema":{"type":"string","enum":["HELPDOG","PARKING","WHEELCHAIR","RESTROOM","AUDIOGUIDE","EXITS"]}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":6}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseDtoListTouristSearchResponseDto"}}}}}}},"/api/v1/search/course":{"get":{"tags":["검색 API"],"summary":"제목기반 코스 검색 API ","description":"검색할 코스 제목을 쿼리파라미터로 넘겨서 코스 검색하는 기능","operationId":"searchCourse","parameters":[{"name":"searchQuery","in":"query","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":12}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CourseResponseDto"}}}}}}}},"/api/v1/notices/{noticeId}":{"get":{"tags":["공지사항 API"],"summary":"개별 공지사항 정보 API ","description":"개별 공지사항 정보 조회 API 입니다.","operationId":"getNoticeDetail","parameters":[{"name":"noticeId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NoticeDetailResponseDto"}}}}}}},"/api/v1/mypage/touristAttraction/likes":{"get":{"tags":["마이페이지 API"],"summary":"내 관광지 좋아요 API","description":"내 관광지 좋아요 리스트 API 입니다.","operationId":"getTouristAttractionLikes","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":6}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TouristSearchResponseDto"}}}}}}}},"/api/v1/mypage/faqs/{faqsId}":{"get":{"tags":["마이페이지 API"],"summary":"개별 자주 묻는 질문 정보 API ","description":"개별 자주 묻는 질문 정보 API입니다.","operationId":"getNoticeDetail_1","parameters":[{"name":"faqsId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FaqDetailResponseDto"}}}}}}},"/api/v1/mypage/course/likes":{"get":{"tags":["마이페이지 API"],"summary":"내 코스 좋아요 API","description":"내 코스 좋아요 리스트 API 입니다.","operationId":"getCourseLikes","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":3}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SearchCourseResponseDto"}}}}}}}},"/api/v1/attraction/{touristAttractionId}/info":{"get":{"tags":["관광지 API"],"summary":"관광지 기본 정보","description":"Path Variable로 받은 관광지의 기본 정보를 응답합니다.","operationId":"getTouristAttractionInfo","parameters":[{"name":"touristAttractionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseDtoAttractionInfoResponseDto"}}}}}}},"/api/v1/attraction/{touristAttractionId}/gallery":{"get":{"tags":["관광지 API"],"summary":"관광지 이미지 정보","description":"Path Variable로 받은 관광지의 이미지 정보를 응답합니다.","operationId":"getTouristAttractionGallery","parameters":[{"name":"touristAttractionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseDtoListGalleryResponseDto"}}}}}}},"/api/v1/attraction/{touristAttractionId}/detail":{"get":{"tags":["관광지 API"],"summary":"관광지 상세 정보","description":"Path Variable로 받은 관광지의 상세 정보를 응답합니다.","operationId":"getTouristAttractionDetail","parameters":[{"name":"touristAttractionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseDtoAttractionDetailResponseDto"}}}}}}},"/api/retry/{areaCode}/{numOfRows}/{pageNo}":{"get":{"tags":["관광지 DB 초기화 API"],"summary":"관광지 에러 재시도 초기화","description":"하드코딩으로 저장 실패한 요소를 다시 시도합니다. 재빌드 하세요.","operationId":"retryByContentId","parameters":[{"name":"areaCode","in":"path","required":true,"schema":{"type":"string"}},{"name":"numOfRows","in":"path","required":true,"schema":{"type":"string"}},{"name":"pageNo","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/perOne/{apiAreaCode}/{numOfRows}/{pageNo}":{"get":{"tags":["관광지 DB 초기화 API"],"summary":"다수의 관광지 초기화","description":"Path Variable로 받은 다수의 관광지의 정보대로 공공API로부터 초기화시킵니다.","operationId":"initDb","parameters":[{"name":"apiAreaCode","in":"path","required":true,"schema":{"type":"string"}},{"name":"numOfRows","in":"path","required":true,"schema":{"type":"string"}},{"name":"pageNo","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/course-generator":{"get":{"tags":["관광지 DB 초기화 API"],"summary":"코스 DB 초기화 API","description":"코스를 관광지 데이터를 기반으로 만들고 Course 테이블의 over_view, distance, course_type, brief_description 데이터를 GPT를 통해 파싱후 자동 저장합니다.","operationId":"coureGen","parameters":[{"name":"areacode","in":"query","required":true,"schema":{"type":"string"}},{"name":"iter","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}},"put":{"tags":["관광지 DB 초기화 API"],"summary":"코스 DB 초기화 API","description":"코스를 관광지 데이터를 기반으로 만들고 Course 테이블의 over_view, distance, course_type, brief_description 데이터를 GPT를 통해 파싱후 자동 저장합니다.","operationId":"coureGen_3","parameters":[{"name":"areacode","in":"query","required":true,"schema":{"type":"string"}},{"name":"iter","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}},"post":{"tags":["관광지 DB 초기화 API"],"summary":"코스 DB 초기화 API","description":"코스를 관광지 데이터를 기반으로 만들고 Course 테이블의 over_view, distance, course_type, brief_description 데이터를 GPT를 통해 파싱후 자동 저장합니다.","operationId":"coureGen_2","parameters":[{"name":"areacode","in":"query","required":true,"schema":{"type":"string"}},{"name":"iter","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}},"delete":{"tags":["관광지 DB 초기화 API"],"summary":"코스 DB 초기화 API","description":"코스를 관광지 데이터를 기반으로 만들고 Course 테이블의 over_view, distance, course_type, brief_description 데이터를 GPT를 통해 파싱후 자동 저장합니다.","operationId":"coureGen_5","parameters":[{"name":"areacode","in":"query","required":true,"schema":{"type":"string"}},{"name":"iter","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}},"options":{"tags":["관광지 DB 초기화 API"],"summary":"코스 DB 초기화 API","description":"코스를 관광지 데이터를 기반으로 만들고 Course 테이블의 over_view, distance, course_type, brief_description 데이터를 GPT를 통해 파싱후 자동 저장합니다.","operationId":"coureGen_6","parameters":[{"name":"areacode","in":"query","required":true,"schema":{"type":"string"}},{"name":"iter","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}},"head":{"tags":["관광지 DB 초기화 API"],"summary":"코스 DB 초기화 API","description":"코스를 관광지 데이터를 기반으로 만들고 Course 테이블의 over_view, distance, course_type, brief_description 데이터를 GPT를 통해 파싱후 자동 저장합니다.","operationId":"coureGen_1","parameters":[{"name":"areacode","in":"query","required":true,"schema":{"type":"string"}},{"name":"iter","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}},"patch":{"tags":["관광지 DB 초기화 API"],"summary":"코스 DB 초기화 API","description":"코스를 관광지 데이터를 기반으로 만들고 Course 테이블의 over_view, distance, course_type, brief_description 데이터를 GPT를 통해 파싱후 자동 저장합니다.","operationId":"coureGen_4","parameters":[{"name":"areacode","in":"query","required":true,"schema":{"type":"string"}},{"name":"iter","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"TouristAttractionLikeResponseDto":{"type":"object","properties":{"likeId":{"type":"integer","format":"int64"},"touristAttractionId":{"type":"integer","format":"int64"},"username":{"type":"string"}}},"SignUpRequestDto":{"required":["email","nickname","password","role"],"type":"object","properties":{"email":{"type":"string"},"nickname":{"type":"string"},"role":{"type":"string"},"password":{"type":"string"}}},"SignUpResponseDto":{"type":"object","properties":{"message":{"type":"string"}}},"NoticeRequestDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"content":{"type":"string"}}},"NoticeResponseDto":{"type":"object","properties":{"noticeId":{"type":"integer","format":"int64"}}},"InquiryRequestDto":{"type":"object","properties":{"title":{"type":"string"},"inquiryType":{"type":"string"},"userId":{"type":"integer","format":"int64"},"content":{"type":"string"}}},"InquiryResponseDto":{"type":"object","properties":{"inquiryId":{"type":"integer","format":"int64"}}},"FaqRequestDto":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"}}},"FaqResponseDto":{"type":"object","properties":{"faqId":{"type":"integer","format":"int64"}}},"TouristAttractionLikeRequestDto":{"type":"object","properties":{"touristAttractionId":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"}}},"CourseLikeRequestDto":{"type":"object","properties":{"courseId":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"}}},"CourseLikeResponseDto":{"type":"object","properties":{"likeId":{"type":"integer","format":"int64"},"courseId":{"type":"integer","format":"int64"},"username":{"type":"string"}}},"ResponseDtoListTouristSearchResponseDto":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/TouristSearchResponseDto"}}}},"TouristSearchResponseDto":{"type":"object","properties":{"title":{"type":"string"},"areaCode":{"type":"string"},"overView":{"type":"string"},"contentId":{"type":"integer","format":"int64"},"contentTypeId":{"type":"string"},"thumbnailUrl":{"type":"string"},"helpdog":{"type":"boolean"},"parking":{"type":"boolean"},"wheelchair":{"type":"boolean"},"restroom":{"type":"boolean"},"audioguide":{"type":"boolean"},"exits":{"type":"string"}}},"CourseResponseDto":{"type":"object","properties":{"courseId":{"type":"integer","format":"int64"},"courseType":{"type":"string"},"title":{"type":"string"},"briefDescription":{"type":"string"},"overView":{"type":"string"},"difficulty":{"type":"string"},"distance":{"type":"number","format":"double"}}},"NoticeDetailResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"content":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"InquiryDetailResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"inquiryType":{"type":"string"},"answered":{"type":"boolean"}}},"FaqDetailResponseDto":{"type":"object","properties":{"faqId":{"type":"integer","format":"int64"},"title":{"type":"string"},"content":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"AttractionBasicResponseDto":{"type":"object","properties":{"title":{"type":"string"},"contentId":{"type":"string"}}},"SearchCourseResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"courseType":{"type":"string"},"title":{"type":"string"},"briefDescription":{"type":"string"},"distance":{"type":"number","format":"double"},"attractions":{"type":"array","items":{"$ref":"#/components/schemas/AttractionBasicResponseDto"}}}},"AttractionInfoResponseDto":{"type":"object","properties":{"contentTypeId":{"type":"string"},"contentId":{"type":"string"},"address1":{"type":"string"},"address2":{"type":"string"},"tel":{"type":"string"},"homepage":{"type":"string"},"useTime":{"type":"string"},"likeCount":{"type":"integer","format":"int64"},"view":{"type":"integer","format":"int64"},"title":{"type":"string"},"helpDog":{"type":"boolean"},"restRoom":{"type":"boolean"},"wheelchair":{"type":"boolean"},"audioGuide":{"type":"boolean"},"parking":{"type":"boolean"}}},"ResponseDtoAttractionInfoResponseDto":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AttractionInfoResponseDto"}}},"GalleryResponseDto":{"type":"object","properties":{"imageUrl":{"type":"string"}}},"ResponseDtoListGalleryResponseDto":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/GalleryResponseDto"}}}},"AttractionDetailResponseDto":{"type":"object","properties":{"contentId":{"type":"integer","format":"int64"},"overView":{"type":"string"},"useFee":{"type":"string"},"brailleBlock":{"type":"boolean"},"elevator":{"type":"boolean"},"guideHuman":{"type":"boolean"},"signGuide":{"type":"boolean"},"stroller":{"type":"boolean"},"lactationroom":{"type":"boolean"}}},"ResponseDtoAttractionDetailResponseDto":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AttractionDetailResponseDto"}}}}}}